requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
Pillow==10.4.0
